refactor: share one index-key delete helper
